move euclid to lib/
