+++ tuxemon/lib/euclid.py
