+++ tuxemon/map.py
@@ -36,7 +36,7 @@ import pyscroll
 
 from tuxemon.compat import Rect
 from tuxemon import prepare
-from tuxemon.euclid import Vector2, Vector3, Point2
+from tuxemon.lib.euclid import Vector2, Vector3, Point2
 from tuxemon.tools import round_to_divisible
 
 logger = logging.getLogger(__name__)
